requirements.txt
django-rest-framework
django-cors-headers
pytils